src/requirements.txt
absl-py==0.6.1
altgraph==0.16.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
biopython==1.73
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info==0.18
evdev==0.7.0
future==0.17.1
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
httplib2==0.9.2
idna==2.6
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
macholib==1.11
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
netifaces==0.10.4
numpy==1.15.4
oauth==1.0.1
olefile==0.45.1
pefile==2018.8.8
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyInstaller==3.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.0
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2018.3
pyudev==0.21.0
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.20.1
scipy==1.1.0
scour==0.36
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sklearn==0.0
system-service==0.3
systemd-python==234
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==0.14.1
xkit==0.0.0
youtube-dl==2018.3.14
zope.interface==4.3.2
